requirements.txt
pyspark
kafka-python
streamlit
pandas
pyyaml